Tests added
