+++ deployment.yaml
@@ -1,4 +1,4 @@
-apiVersion: v1
+wapiVersion: v1
 kind: Service
 metadata:
   name: package-json-image
