+++ package.json
@@ -32,5 +32,6 @@
     "eslint-config-next": "^16.0.1",
     "jest": "^30.2.0",
     "ts-jest": "^29.4.6"
-  }
+  },
+  "type": "module"
 }
