+++ src/components/atoms/fixtures-image/fixtures-image.js
@@ -3,13 +3,13 @@ import styles from "./fixtures-image.module.css";
 const FixturesImage = ({ img, children }) => {
   return (
     <div className={styles.wrapper}> 
-    <section 
-      className={styles.tableFixtures} 
-      style={{
-        backgroundImage: `url(${img.src})`,
-      }}>
-        {children}
-      </section>
+      <div
+        className={styles.tableFixtures} 
+        style={{
+          backgroundImage: `url(${img.src})`,
+        }}>
+          {children}
+        </div>
     </div>
   );
 };
