+++ src/components/organisms/calendar-dates/index.js
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { Months } from "../../atoms/months";
+import { Months } from "../months";
 import styles from "./calendar-dates.module.css";
 import { CalendarArrow } from "@/components/molecules/calendar-arrow";
