+++ src/components/organisms/months/__tests__/index.test.js
@@ -1,0 +1,8 @@
+import { render } from "@testing-library/react"
+
+
+describe('Months', () => {
+  it('render on page', () => {
+    render
+  })
+})
