+++ src/components/organisms/months/months.module.css
@@ -10,6 +10,7 @@
 
 .container > * {
   scroll-snap-align: center;
+  list-style: none;
 }
 
 .container > *:hover {
@@ -21,6 +22,20 @@
   font-family: Helvetica;
   transform: scale(1);
   animation: bob 1.8s ease-in-out infinite alternate;
+}
+
+.section > button {
+  appearance: none;
+  -webkit-appearance: none;
+  background: none;
+  border: 0;
+  padding: 0;
+  margin: 0;
+  font: inherit;
+  color: inherit;
+  text-align: inherit;
+  line-height: inherit;
+  letter-spacing: inherit;
 }
 
 .opacity100 {
@@ -60,6 +75,7 @@
     transition: opacity 0.5s ease-in-out;
     height: 14rem;
     animation: fade-in 0.5s ease-out;
+    margin-left: 1rem;
   }
 
   .section {
